Add setup process.
If no minions have any role assigned, we will redirect to `/setup` where we will
wait for user input for certain settings.

`/setup` cannot be visited if there is any minion with an assigned role.

We will also run the discovery, assign roles and run the orchestration.

The root path `/` will contain a basic monitoring tool.