Ensure that gems are the same in development/test and production
By removing code conditionals on the Gemfile we ensure that a restriction
that (dis)appears in production don't shift dependencies versions. Is one
of the good points of bundler, that we can use groups to avoid installing
unnecessary gems in certain environments, but the versions of the gems
are the same both in development/test and production.

This makes bug easier to reproduce if they are related to dependencies.
We can still use groups to only install the groups that we want, or
discard groups that we won't need (--without development test in production
for instance). But we are sure that the installed gems are always the same
versions in all environments.