Serve the autoyast content even when a SCC/SMT connection error happens
The service unavailable http error code (503) will only be returned now
until the first page of the setup is completed (so we have the dashboard
address to serve the autoyast profile).

When one of the following exceptions happens:
  - Missing registration code
  - Missing credentials for SCC/SMT
  - Any HTTP error while connecting to SCC/SMT

We will still serve the autoyast content with a comment explaining why
the system cannot be registered. The returned http error code will be
success (200), but we will show a dialog stating the problem that was
detected and why it won't be registered.

Fixes: bsc#1049948