Remove unneeded unprotected pillars
There's no possible use case for these pillars in which they were
filled with information and then the customer will clear them (as they
are required).

Keep things as simple as possible until we refactor this part.

Without this change, a specially crafted HTTP request could be made to
remove the content of those pillar values, when this is not ideal.