If a system won't be registered, set a default timeout for that warning during the installation.
Right now it's completely necessary for the customer to confirm the "this
system is not going to be registered" warning, or the installation won't
continue.

Make this dialog have a timeout so we show it at the very beginning, but
it's also possible to install unattended workers without further action.

Fixes: bsc#1059016