When we are back in Velum from Dex, allow for assets to be loaded
If we forced the location directly, javascript assets were never loaded
as their loading are interrupted by the browser because of the `href` change.

This ensures we first load all the assets (so delete, put, patch links work
through jquery as usual), and then we force the redirection to happen.

Fixes: bsc#1066611